#22 Simplify feature names
