--- lib/features/book_storage/domain/repositories/book_repository.dart
@@ -1,12 +1,0 @@
-import 'package:rust_core/result.dart';
-import 'package:storyscape/features/book_storage/domain/entities/existing_book.dart';
-import 'package:storyscape/features/book_storage/domain/entities/new_book.dart';
-import 'package:storyscape/features/book_storage/domain/entities/stored_book.dart';
-
-abstract interface class BookRepository {
-  Future<Result<ExistingBook, String>> storeNewBook(NewBook book);
-
-  Future<Result<ExistingBook, String>> retrieveBookById(int id);
-
-  FutureResult<List<StoredBook>, String> fetchAllBooks();
-}
