+++ lib/features/new_book/domain/entities/existing_book.dart
