+++ lib/features/new_book/domain/entities/new_book.dart
