+++ lib/features/new_book/domain/entities/stored_book.dart
