+++ lib/features/select_book/ui/cubit/book_selection_state.dart
